third_party/WebKit/LayoutTests/fast/css/invalidation/shadow-boundary-crossing-expected.txt

Invalidation sets should not apply across shadow roots if tree boundary crossing rules are not used.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.updateStyleAndReturnAffectedElementCount() is 11
PASS internals.updateStyleAndReturnAffectedElementCount() is 31
PASS successfullyParsed is true

TEST COMPLETE
